+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - GetEvents inline docs
+- GraphQL schema removed Parent from coin_owner enumeration in StandardCoin fragment
+- GraphQL schema removed Parent from object_owner enumeration in StandardObject fragment
+- GraphQL schema removed event type from event enumeration in StandardEvent fragment
 
 ### Removed